# spec/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.19.2-SNAPSHOT
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local API server
paths:
  /namespaces/{namespace}:
    parameters:
      - $ref: '#/components/parameters/namespace'
    put:
      summary: Create a namespace
      description: Creates a new namespace object. A namespace enables the contextual grouping of related
        jobs and datasets. Namespaces must contain only letters (`a-z`, `A-Z`), numbers (`0-9`),
        underscores (`_`), dashes (`-`), colons (`:`), slashes (`/`), or dots (`.`).
        A namespace is case-insensitive with a maximum length of `1024` characters.
        Note jobs and datasets will be unique within a namespace, but not across namespaces.
      tags:
        - Namespaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedNamespace'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
    get:
      summary: Retrieve a namespace
      description: Returns a namespace.
      tags:
        - Namespaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
  /namespaces:
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all namespaces
      description: Returns a list of namespaces.
      tags:
        - Namespaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceList'
  /sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /sources:
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all sources
      description: Returns a list of sources.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
...
  /namespaces/{namespace}/jobs/{job}/runs:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    post:
      summary: Create a run
      description: Creates a new run object for a job.
      deprecated: true
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedRun'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all runs
      description: Returns a list of runs for a job.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
  /jobs/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/runId'
    get:
      summary: Retrieve a run
      description: Retrieve a run.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteRun'
  /jobs/runs/{id}/start:
    parameters:
      - $ref: '#/components/parameters/runId'
      - $ref: '#/components/parameters/at'
    post:
      summary: Start a run
      description: Marks the run as `RUNNING`.
      deprecated: true
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteRun'
  /jobs/runs/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/runId'
      - $ref: '#/components/parameters/at'
    post:
      summary: Complete a run
      description: Marks the run as `COMPLETED`.
      deprecated: true
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
  /jobs/runs/{id}/fail:
    parameters:
      - $ref: '#/components/parameters/runId'
      - $ref: '#/components/parameters/at'
    post:
      summary: Fail a run
      description: Marks the run as `FAILED`.
      deprecated: true
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteRun'
  /jobs/runs/{id}/abort:
    parameters:
      - $ref: '#/components/parameters/runId'
      - $ref: '#/components/parameters/at'
    post:
      summary: Abort a run
      description: Marks the run as `ABORTED`.
      deprecated: true
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteRun'
...
components:
  parameters:
    at:
      name: at
      in: query
      description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the time when the run transitioned.
      required: false
      schema:
        type: string
        format: date-time
    namespace:
      name: namespace
      in: path
      description: The name of the namespace.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-namespace
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
...
    job:
      name: job
      in: path
      description: The name of the job.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-job
    runId:
      name: id
      in: path
      description: The ID of the run.
      required: true
      schema:
        type: string
        format: uuid
        example: ea9badc5-7cb2-49af-9a9f-155771d3a797
...
    limit:
      name: limit
      in: query
      description: The number of results to return from offset
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results
      required: false
      schema:
        type: integer
        default: 0
...
  schemas:
    CreatedNamespace:
      type: object
      properties:
        ownerName:
          description: The owner of the namespace.
          type: string
        description:
          description: The description of the namespace.
          type: string
      required:
        - ownerName
      example:
        ownerName: me
        description: My first namespace!
    Namespace:
      type: object
      properties:
        name:
          description: The name of the namespace.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the namespace was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the namespace was updated.
          type: string
          format: date-time
        ownerName:
          description: The owner of the namespace.
          type: string
        description:
          description: The description of the namespace.
          type: string
      example:
        name: my-namespace
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        ownerName: me
        description: My first namespace!
    NamespaceList:
      type: object
      properties:
        namespaces:
          type: array
          items:
            $ref: '#/components/schemas/Namespace'
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    SourceList:
      type: object
      properties:
        sources:
          items:
            $ref: '#/components/schemas/Source'
...
    DatasetVersionId:
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the dataset.
        name:
          type: string
          description: The name of the dataset.
        version:
          type: string
          format: uuid
          description: The version of the dataset.
      required:
        - namespace
        - name
        - version
...
    JobVersionId:
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the job.
        name:
          type: string
          description: The name of the job.
        version:
          type: string
          format: uuid
          description: The version of the job.
...
    CreatedRun:
      type: object
      properties:
        id:
          description: An _optional_ user-provided unique ID of the run. A run ID **must** be an [UUID](https://tools.ietf.org/html/rfc4122).
                       If an ID for the run is not provided, a random UUID will be generated for the given run.
          type: string
          format: uuid
        nominalStartTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal start time of the run.
          type: string
          format: date-time
        nominalEndTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal end time of the run.
          type: string
          format: date-time
        args:
          description: The arguments of the run.
          type: object
      example:
        args: {'email': 'me@example.com', 'emailOnFailure': false, 'emailOnRetry': true, 'retries': 1}
    IncompleteRun:
      type: object
      properties:
        id:
          description: The ID of the run.
          type: string
          format: uuid
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the run was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the run was updated.
          type: string
          format: date-time
        nominalStartTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal start time of the run.
          type: string
          format: date-time
        nominalEndTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal end time of the run.
          type: string
          format: date-time
        state:
          description: The current state of the run.
          type: string
          enum: [NEW, RUNNING, COMPLETED, FAILED, ABORTED]
        startedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the start time of the run.
          type: string
          format: date-time
        endedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the end time of the run.
          type: string
          format: date-time
        durationMs:
          type: integer
          description: The total duration of the run. Measured in milliseconds.
        args:
          description: The arguments of the run.
          type: object
        facets:
          $ref: '#/components/schemas/RunFacets'
      example:
        id: 870492da-ecfb-4be0-91b9-9a89ddd3db90
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        nominalStartTime: null
        nominalEndTime: null
        state: RUNNING
        startedAt: 2019-05-09T15:17:32.690346
        endedAt: null
        durationMs: null
        args: {'email': 'me@example.com', 'emailOnFailure': 'false', 'emailOnRetry': 'true', 'retries': '1'}
        facets: {}
    Run:
      type: object
      allOf:
        - $ref: '#/components/schemas/IncompleteRun'
        - type: object
          properties:
            jobVersion:
              $ref: '#/components/schemas/JobVersionId'
            inputVersions:
              type: array
              items:
                $ref: '#/components/schemas/DatasetVersionId'
            outputVersions:
              type: array
              items:
                $ref: '#/components/schemas/DatasetVersionId'
            facets:
              $ref: '#/components/schemas/RunFacets'
      example:
        id: 870492da-ecfb-4be0-91b9-9a89ddd3db90
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        nominalStartTime: null
        nominalEndTime: null
        state: COMPLETED
        startedAt: 2019-05-09T15:17:32.690346
        endedAt: 2019-05-09T20:05:46.815920Z
        durationMs: 4250894125
        args: {'email': 'me@example.com', 'emailOnFailure': 'false', 'emailOnRetry': 'true', 'retries': '1'}
        facets: {}
    RunList:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
...
    RunFacets:
      description: The facets of the run. For a full list of standard run facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    CustomFacet:
      description: A custom facet enables the extension of _dataset_, _job_, and _run_ metadata. A custom facet **must** also have a schema,
                   where a version of the schema is identifiable via a URL. A field within the schema **must** not start with an underscore (`_`).
      allOf:
        - $ref: '#/components/schemas/BaseFacet'
        - type: object
          additionalProperties: true
    BaseFacet:
      description: All base facet fields are prefixed with an underscore (`_`) to avoid field naming conflicts defined with other facets.
      type: object
      properties:
        _producer:
          description: URI identifying the producer of this metadata. For example this could be a git url with a given tag or sha
          type: string
          format: uri
          example: https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client
        _schemaURL:
          description: The URL to the corresponding version of the schema definition following a $ref URL Reference (see https://swagger.io/docs/specification/using-ref/)
          type: string
          format: uri
          example: https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/spec/OpenLineage.yml#MyCustomJobFacet
      required:
        - _producer
        - _schemaURL
...